Remove unncessary input param of key pair type for ED25519 sign and verify.
This input isn't needed since there is only one ED25519 key pair algo type.  This is unlike RSA or ECDSA where there is differences in algo type because of bitsize or curve.  Removing this param for ED25519 cleans up functions a bit since we don't gather a needless parameter.

As part of this, caught a bug where a non-matching algo type could be provided and used for hashing.  This was wrong, we should use the correct hashing function based on the key pair algo.  Implement tooling to check this.