Define new/improved functionality for checking if a license file is expired. This will result in some breakage in some usage.
Verification of a license can now (optionally based on func used) check if a license is expired.  This alleviates the need to separately check if the file's signature is valid and then check if the expiration date in the license file is in the past.  Just a bit simpler for most uses.

The reverify functionality was removed since it was rarely, if ever, used.  Instead, simply re-read the license file and verify it again.  That is all really all that the reverifiation func did, it just did it all internally.