Revise function naming for handling signature verification and explain why verification of signature and checking of expiration date are done separately.
Previously, the funcs used to verify a license key file's signature were named Verify() or Verify{RSA,ED25519,ECDSA}().  Now, these are named VerifySignature() or VerifySignature{RSA,ED25519,ECDSA}().  This was done to more clearly state that these funcs only verify a license key file's signature, nothing else (expiration is not validated).  This was also clarified in comments explaining that these funcs to do not check if the license key file has expired.  The old Verify() funcs still remain, but simply call VerifySignature() and are deprecated.

The contents of the licensefile.go file was rearranged a bit to be more clear and organized about use; funcs are grouped for use when writing a license key file or reading/validating a license key file.