Fix bug where if an numeric value was provided for a text custom field, the license could not be created.
This was caused by using "v-model.number" on a field that expected textual input.  If the input was text, as the input allows, all was fine.  However, if a user typed in a number, Vue would convert the input to a JS number, send the data to the server as such, and the server would try to parse the number (float64 on Go) to a null.String which doens't work.