Set SameSite filed for login and 2FA tokens from Strict to Lax to allow navigating from browser history in Chrome.
In Chrome, if a user was logged in, opened history and clicked on link to navigate to a page from the history, the cookies would not be set and therefore the user would be shown an error and redirected to login page.  This was caused because Chrome doesn't send Strict cookies when navigating from history (Firefox does).  Setting cookies to Lax fixes this.