Rebuild build script to use config file instead of hardcoded config.
This was mostly done to allow the build script to be reused for other apps/projects easily without having to modify the underlying build script code. Therefore, if changes were made to build.go, the updated file could be easily compared against build.go in another app/project's repo and we wouldn't see diff lines for the config.