Improve handling of API example for creating a license by showing button to update example request based on actual data in form.
Previously, the API request was more of just an example, based on license data in form at the time the example request was shown.  The example request was never updated.  Now, there is a button to refresh the example request that causes it to use the actual, real, live data from the form.

It would be nice if the example request updated automatically, but I couldn't get Vue's 'watch' to work and I didn't want to call a v-on:change method on every <input> that could affect the example request.

Also updated example request to use Authorization header to provide API key, not providing API key in a -d data field.