More Vue 3 and Bootstrap 5 rewriting.
Got all the Vue 2 -> 3 rewritten.  Updated a lot of JS just for consistency and clarity purposes.  Lots of testing to do.

Got a lot of the Bootstrap 5 stuff fixed.  Still lots of testing and fixes to implement.

Worked on updating GUI a bit (btn-block, page-header dropdown menu for "home" and user settings.