Ripped out YAML and replaced with JSON since JSON is part of the standard library and YAML isn't.
The YAML library we were using is archived now too.  It just makes sense to use std lib stuff since there is a higher chance of support/fixes.  Plus, JSON is a bit easier to work with in regards to how strings are displayed (always encapsulated in double-quotes) and I like how arrays and objects are displayed.