Lots of sign/verify work.
Rewrote how public and private keys are encoded to just encode them in hex, not PEM format.  This was to make using online decoding tools easier.  Also could make end-usage easier since then we don't have to explain decoding of public key .....Not sure if this was the correct decision, might revert back to PEM encoding and just have to explain decoding better.